requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pytesseract==0.3.10
paddleocr==2.6.1.3
easyocr==1.7.1
PyMuPDF==1.23.26
pdfplumber==0.11.4
camelot-py[cv]==0.11.0
pdfminer.six==20240706
python-docx==1.1.2
reportlab==4.2.5
fpdf==1.7.2
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
opencv-python==4.10.0.84
Pillow==10.3.0
python-multipart==0.0.9
